+++ MAT401_Project/MAT401_Project/Runge_Kutta.h
@@ -8,15 +8,15 @@ public:
 	Runge_Kutta();
 	~Runge_Kutta();
 
-	std::vector<std::vector<float>> RK2(float x0, float y0, float xe, float dx);
-	std::vector<std::vector<float>> RK4(float x0, float y0, float xe, float dx);
+	std::vector<std::vector<double>> RK2(double x0, double y0, double xe, double dx);
+	std::vector<std::vector<double>> RK4(double x0, double y0, double xe, double dx);
 private:
 	math_library* maths_;
 
-	float inline function(float x, float y) {return  5.0f * (pow(x,4)); }
+	double inline function(double x, double y) {return  5.0f * (pow(x,4)); }
 
-	float inline function2(float TE, float T) { 
-		float i = TE - T;
+	double inline function2(double TE, double T) { 
+		double i = TE - T;
 		return  2*i; 
 		}
 };
