+++ MAT401_Project/MAT401_Project/Task_3_and_4.h
@@ -8,15 +8,15 @@ public:
 	Task_3_and_4();
 	~Task_3_and_4();
 
-	void Init(std::array<float,3> Initial_velocity, std::array<float, 3> Initial_position);
-	void Solve_Task_3(std::array<float, 3> Initial_velocity, std::array<float, 3> Initial_position,std::vector<std::vector<float>>& Out_Vx, std::vector<std::vector<float>>& Out_Vy, std::vector<std::vector<float>>& Out_Vz, std::vector<std::vector<float>>& Out_Px, std::vector<std::vector<float>>& Out_Py, std::vector<std::vector<float>>& Out_Pz, float Step, float Start, float End);
+	void Init(std::array<double,3> Initial_velocity, std::array<double, 3> Initial_position);
+	void Solve_Task_3(std::array<double, 3> Initial_velocity, std::array<double, 3> Initial_position,std::vector<std::vector<double>>& Out_Vx, std::vector<std::vector<double>>& Out_Vy, std::vector<std::vector<double>>& Out_Vz, std::vector<std::vector<double>>& Out_Px, std::vector<std::vector<double>>& Out_Py, std::vector<std::vector<double>>& Out_Pz, double Step, double Start, double End);
 
-	float Semi_Implict_Euler(float Initial_Value,float Diffrence,float Step);
+	double Semi_Implict_Euler(double Initial_Value, double Diffrence, double Step);
 
 private:
-	std::array<std::vector<float>, 3> velocity;
-	std::array<std::vector<float>, 3> position;
-	const float Gravity = -9.8f;
+	std::array<std::vector<double>, 3> velocity;
+	std::array<std::vector<double>, 3> position;
+	const double Gravity = -9.8f;
 	math_library* maths_;
 };
