+++ MAT401_Project/MAT401_Project/math_library.cpp
@@ -8,9 +8,9 @@ math_library::~math_library()
 {
 }
 
-std::vector<float> math_library::Generate_Half_open_interval(float step, float start, float end)
+std::vector<double> math_library::Generate_Half_open_interval(double step, double start, double end)
 {
-	std::vector<float> output;
+	std::vector<double> output;
 	for (float i = start; i < end+step; i+=step)
 	{
 		////std::cout << "Generating values :" << i << "\n";
@@ -19,9 +19,9 @@ std::vector<float> math_library::Generate_Half_open_interval(float step, float s
 	return output;
 }
 
-std::vector<float> math_library::Generate_zeros(float length)
+std::vector<double> math_library::Generate_zeros(double length)
 {
-	std::vector<float> output;
+	std::vector<double> output;
 	for (float i = 1; i <= length; i++)
 	{
 		//std::cout << "Generating values :" << i << "\n";
