+++ MAT401_Project/MAT401_Project/math_library.h
@@ -6,8 +6,8 @@ public:
 	math_library();
 	~math_library();
 
-	std::vector<float> Generate_Half_open_interval(float step, float start,float end);
+	std::vector<double> Generate_Half_open_interval(double step, double start, double end);
 
-	std::vector<float> Generate_zeros(float length);
+	std::vector<double> Generate_zeros(double length);
 };
